requirements.txt
annotated-types==0.7.0
anthropic==0.93.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.9.0
PyJWT==2.12.1
PyNaCl==1.6.2
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
